Get rid of typedefs/SignalInfo and replace AudioMetaData
