+++ torchaudio/csrc/sox_io.h
@@ -3,12 +3,25 @@
 
 #include <torch/script.h>
 #include <torchaudio/csrc/sox_utils.h>
-#include <torchaudio/csrc/typedefs.h>
 
 namespace torchaudio {
 namespace sox_io {
 
-c10::intrusive_ptr<torchaudio::SignalInfo> get_info(const std::string& path);
+struct SignalInfo : torch::CustomClassHolder {
+  int64_t sample_rate;
+  int64_t num_channels;
+  int64_t num_frames;
+
+  SignalInfo(
+      const int64_t sample_rate_,
+      const int64_t num_channels_,
+      const int64_t num_frames_);
+  int64_t getSampleRate() const;
+  int64_t getNumChannels() const;
+  int64_t getNumFrames() const;
+};
+
+c10::intrusive_ptr<SignalInfo> get_info(const std::string& path);
 
 c10::intrusive_ptr<torchaudio::sox_utils::TensorSignal> load_audio_file(
     const std::string& path,
